Add `overlay_0_16` to outputs and use it in CI for `run executable on zls`
Prior to this change, this CI step was run on `default_0_13`. This does not work with our
build system and code changes of a previous commit any more.

This change adds the `overlay_0_16` and uses it for the CI job. For this, we force the linkage
to be static becuase if we don't there is an error with dynamic linking.